move custom EventGeneratorWrapper to own module
