+++ webhooks/dependencies/sse_manager.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 import asyncio
 import logging
 import time
@@ -9,6 +7,7 @@ from typing import Any, AsyncGenerator, Tuple
 
 from shared import TopicItem
 from shared.models.topics import WEBHOOK_TOPIC_ALL
+from webhooks.dependencies.event_generator_wrapper import EventGeneratorWrapper
 from webhooks.dependencies.service import Service
 
 LOGGER = logging.getLogger(__name__)
@@ -155,7 +154,7 @@ class SseManager:
                     stop_event.set()
 
         return EventGeneratorWrapper(
-            event_generator(), self, wallet, topic, populate_task
+            generator=event_generator(), populate_task=populate_task
         )
 
     async def _populate_client_queue(
@@ -265,28 +264,3 @@ async def _copy_queue(
     LOGGER.debug("SSE Manager: Finished repopulating cache")
 
     return lifo_queue, fifo_queue
-
-
-class EventGeneratorWrapper:
-    def __init__(
-        self,
-        generator: AsyncGenerator[TopicItem, Any],
-        sse_manager: SseManager,
-        wallet: str,
-        topic: str,
-        populate_task: asyncio.Task,
-    ):
-        self.generator = generator
-        self.sse_manager = sse_manager
-        self.wallet = wallet
-        self.topic = topic
-        self.populate_task = populate_task
-
-    async def __aenter__(self):
-        return self.generator
-
-    async def __aexit__(self, exc_type, exc_val, exc_tb):
-        await self.generator.aclose()  # Close the async generator
-        await self.sse_manager.cleanup_task(
-            self.wallet, self.topic, self.populate_task
-        )  # Call the cleanup method on the SseManager instance
