+++ webhooks/routers/sse.py
@@ -8,7 +8,8 @@ from sse_starlette.sse import EventSourceResponse
 
 from shared import WEBHOOK_TOPIC_ALL, APIRouter
 from webhooks.dependencies.container import Container
-from webhooks.dependencies.sse_manager import EventGeneratorWrapper, SseManager
+from webhooks.dependencies.event_generator_wrapper import EventGeneratorWrapper
+from webhooks.dependencies.sse_manager import SseManager
 
 LOGGER = logging.getLogger(__name__)
